requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.0

# Data processing
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.10.0

# Transformers and NLP
transformers>=4.30.0
tokenizers>=0.13.0
sentencepiece>=0.1.99

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Evaluation
scikit-learn>=1.2.0

# Development dependencies (optional)
# Install with: pip install -e ".[dev]"
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
pylint>=2.17.0
mypy>=1.3.0
pre-commit>=3.3.0

# TensorBoard (optional)
TensorBoard>=2.12.0
tensorboardX>=2.6.0

# Documentation (optional)
# Install with: pip install -e ".[docs]"
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0
sphinx-autodoc-typehints>=1.23.0